startet to implement Server
