+++ app/src/main/java/com/example/server/MainActivity.java
@@ -25,45 +25,9 @@ public class MainActivity extends AppCompatActivity {
         findViewById(R.id.btnStartServer).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                System.out.println("yeah");
-                MainActivity server = new MainActivity();
+                MyFirstServer server = new MyFirstServer();
                 server.startListening();
             }
         });
-    }
-    public MainActivity(){
-    }
-
-    private void startListening() {
-        System.out.println("[Server] starten...");
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                while(true) {
-                    try {
-                        ServerSocket serverSocket = new ServerSocket(7000);
-                        System.out.println("[Server] Warten auf Verbindung....");
-                        Socket remoteClientSocket = serverSocket.accept();
-                        System.out.println("[Server] Client verbunden: " + remoteClientSocket.getRemoteSocketAddress());
-
-                        Scanner s = new Scanner(new BufferedReader(new InputStreamReader(remoteClientSocket.getInputStream())));
-                        if (s.hasNextLine()) {
-                            System.out.println("[Server] Message from client: " + s.nextLine());
-                        }
-
-                        PrintWriter pw = new PrintWriter(new OutputStreamWriter(remoteClientSocket.getOutputStream()));
-                        pw.println("Hi Client");
-                        pw.flush();
-                        // Verbindung schließen
-                        s.close();
-                        remoteClientSocket.close();
-                        serverSocket.close();
-
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
-        }).start();
     }
 }
